+++ src/main/scala/com/github/gchudnov/mtg/diagram/Label.scala
@@ -1,0 +1,12 @@
+package com.github.gchudnov.mtg.diagram
+
+/**
+ * Label
+ */
+final case class Label(pos: Int, value: String):
+  def size: Int =
+    value.size
+
+object Label:
+  val empty: Label =
+    Label(0, "")
