+++ src/main/scala/com/github/gchudnov/mtg/diagram/Tick.scala
@@ -1,0 +1,6 @@
+package com.github.gchudnov.mtg.diagram
+
+/**
+ * Tick
+ */
+final case class Tick(pos: Int)
